feat(backend): implement users/followers and users/following
