+++ backend/src/main/java/com/melodify/melodify/model/response/UserResponse.java
@@ -1,0 +1,12 @@
+package com.melodify.melodify.model.response;
+
+import lombok.Data;
+
+@Data
+public class UserResponse {
+    private String username;
+
+    public UserResponse(String username) {
+        this.username = username;
+    }
+}
